[#2516] Document cache arguments and explicitly declare cache options
This commit makes the documentation and typing around the utility
cache decorator a bit more explicit. It also removes generic
kwarg cache option in favour of the explicit timeout argument.
The only other option accepted by the cache setter, other than
`timeout`, is `version`, which we should avoid until there is an
explicit reason to do so to keep the interface explicit.