[#2560] Treat catalogus as a required field on zaaktype response
Previously we allowed for two edge-cases in resolving zaaktypes:

1. We allowed for catalogus to be absent on a ZaakType to
   accommodate eSuite, in which the field was absent.
2. We allowed for importing a ZaakType if it included a URL
   pointing to a Catalogus that had not previously been
   created.

This commit no longer allows for these exceptions. With regards
to 1, eSuite has since implemented changes to include the
catalogus field. As for 2: it was already an edge-case we
felt should raise rather than be suppressed, but this is
especially true now as we're moving to support multiple
ZGW backends. In that scenario, we'll clearly want to
only include ZaakTypes from a known Catalogus API, to avoid
cross-backend ambiguity.