[#2566] Persist originating service on zgw clients
To support multiple zgw backends, we frequently need to determine
which backend to query for which resource. As a first step to
keeping better track of where ZGW objects come from, in this
commit we persist the zgw service used to configure a client
on the client object, so that we can later fetch and persist
this service to facilitate easier target backend resolution.