[#2589] Run CI tests in parallel
We have made several previous attempts at running the CI tests in
parallel, which were ultimately reverted because we ran into the
issue described at https://github.com/django/django/pull/17650 and
https://code.djangoproject.com/ticket/32114 (in this repo, see PRs
1050, 948 and 690.

The Django ticket has been fixed, but will will only be included
in 5.1 onwards, and won't be backported to 4.2.x. However, it seems
the current testsuite no longer fails with the issue reported in the
Django ticket (pickle related errors). It is possible the potentially
failing tests are flaky, but so far we've not run into it during a
handful of parallel runs. This commit will re-enable parallel tests,
and we'll have to see if we run into the same issues again.