Hack around app label rename
Probably the shadiest/trickest part of this upgrade - we have to keep
existing migration history in mind to avoid running them twice (as
those would crash out), but we must assume this SQL will also run
when the new dependencies are set up and that would incorrectly mark
the external package migrations as belonging to the legacy app.

Therefore, we strictly limit the migrations that may be renamed to the
ones that are known - the timestamps added by django definitely help
here.

The digid-eherkenning.oidc migrations also start with a squashed
migration, so there is no collision on the 0001_initial migration :tada: