:alien: Handle charfield -> claimfield changes in mozilla-django-oidc-db
Before 0.16.0, dots in claims were used to split the
bits of the claim to support nested claims, but that
didn't allow claims with a dot in the key to exist/
be supported. This was addressed in the upstream
library - now an array of strings is stored to specify
the claim path to consume.

However, this means that the legacy format of
Mapping[str, str] needs to be adapted to
Mapping[str, list[str]] instead.