:alien: Address broken imports / removed bits
The caching and config mixins were removed, which caused broken imports
and thus 'manage.py check' to fail. A bunch of constructs are no longer
necessary to do overrides, and have been removed or replaced with the
new way to do things - as an in-between step, since most of this will
eventually be sorted out when django-digid-eherkenning is updated with
the OIDC subpackage that handles all this already.

Tests are expected to be broken - most notably the callback flow/
customization mechanism currently is not functional.