:alien: Re-organize callback views
Callbacks now should all go through the same actual callback
view, which dispatches a callback view on a per-config
basis. This allows us to still apply config-specific error
message mappings, but the URL routing is a bit different
because of it.

This feature of the upstream library makes the different
endpoints/redirect-uri's per OIDC flow obsolete - all of
them can use the same endpoint in the future.