:alien: Fixing eherkenning tests/implementation
The extra callback view override in the accounts app can (and should)
be loaded via the config model. Currently this is messy, but will be
cleaned up via proxy models once the vendored
digid_eherkenning_oidc_generics app is gone.

The URL conf overrides are not necessary anymore, because the callback
view handling is done via config model now, so this simplifies the setup
a little bit too. It was a bit of a mindfuck to figure out what is
going on, with three views hooked up to the same url route.