[dimpact-105] Fix ctypes dependencies for admin index fixture
The django-admin-index fixture requires all contenttypes to be
present before being loaded. Previously, we would only ensure
the contenttypes for our own apps were loaded, but this is not
sufficient, as the fixture frequently makes references to third
party apps. This commit clarifies this behavior and ensures that
all contenttypes are loaded.