:whale: Updated Dockerfile, separated Celery containers
- Added os upgrade to be sure the libraries as up to date as possible everytime the image is built.
- Replaced vim with nano
- Used uv instead of pip (super fast drop in replacement)
- Compile translatable messages during build
- Added proper release scheme attributes
- Separated Celery into worker, beat and monitor containers
- Added additional health checks to steer the ordering of container-startup

There was 1 container reserved for all Celery related things:
worker, beat and monitoring. According to Celery documentation
it's best to have a separate Beat process so you can scale
separately. Also, combining monitoring with the events itself
could cause issues. Everything is separated now.

Additional health checks and dependancies are introduced to
Docker-compose so Celery nicely waits for the database to be
prepped.