[#2636] stop using human-friendly names for tasks
The Celery docs make clear that the name parameter of the
task decorator should be unique and that generally it functions
as an identifier the task. From this we can also infer that
names that change at runtime, due to gettext, are undesirable.
This in fact created issue with triggering certain periodic tasks
in the admin, due to a mismatch between the registered name (English)
and the translated name used at runtime (Dutch).

TLDR: let Celery generate the name, and directly translate the periodic
task names for Beat so that the entries in the Django periodic task
admin are human-readable. This is not ideal as we would like to use
(lazy) gettext here, but Celery doesn't allow that, so this strikes
the balance between human-readability and predictable task discovery.