avoid resolving zgw objects in contactmomenten client
The immediate trigger for this is that passing a zaken_client
to the objectcontactmomenten is fraught now we support multiple
zgw backends: which backend do we pass to resolve the zaak?

More generally: resolving does not belong in the API client, which
should focus on fetching data from a service and returning it to
the caller, which is a distinct concern from enriching the objects
thus retrieved.