[#2648] handle legacy case detail URLs with redirect
The case detail URL now contains a reference to the
ZGW api group to be used to fetch the case. It is quite
likely there are various hard-coded URLs in the wild
(especially in notification emails) that point to the
old case detail URLs. This commit adds a redirect
handler for the now-deprecated case detail URL and
tries to handle the request gracefully, depending on
the case.