:card_file_box: [#2617] Set up vendored copy of digid_eh_generics_oidc for migration
The app_label is changed to avoid conflicts with the app_label defined
in the 'digid_eherkenning.oidc' package. Migrations are added to
rename any existing database tables and keep a consistent migration
history/state.

Special care is needed with regards to the django_migrations table -
fixes here cannot be done through migrations, so the AppConfig.ready
hook is used to bring these in a consistent state.

Finally, the code that is replaced by the library is deleted to
de-clutter, which includes uninstalling the models. The migrations are
defined such that the new tables are created first and configuration can
be copied over, which prevents data loss.