:card_file_box: [#2617] Update to proxy models for DigiD/eHerkenning config
Using proxy models is the preferred approach to modify python behaviour
as no database schema changes are usually required. This does require
all code to call the proxy model rather than the concrete base model,
or different outcomes for the same checks would cause problems.

This includes the copy of the legacy data into the new models, which
should be a smooth automatic migration.