:recycle: [#2617] Refactor the OIDC authentication flow
The OIDC libraries support different OIDC-configurations/flow through
the same views and backends, which is what this changeset achieves.

The init views are exposed via separate URLs/endpoints, and bind the
specific OIDC config flavour to be used to the endpoint. This
information is recorded in the (session) state, and consulted again
in the callback flow.

All return flows (when returning from the OIDC Provider) go through
the same callback view, which can route to specific handlers based on
the configuration used. DigiD/eHerkenning is sent to a single
authentication backend, which implements behaviours depending on the
specific flavour used. The admin OIDC configuration is bound to the
admin user backend, isolated from the DigiD/eHerkenning backend.

Currently the callback URLs are still different urls for admin, digid
and eherkenning, but there is no need for this - you can already
use /oidc/callback/ instead (see the setting OIDC_CALLBACK_CLASS).