:boom: [#2617] Update changed setting names for setup-configuration
The configuration settings are tightly coupled to the django model
fields (names and data types). The field names for the configuration
have been changed compared to the vendored library, and the structure
has been changed to be able to use claims with dots in their name.

Unfortunately, this means that breaking changes in the infrastructure/
configuration setup are required because of the way this config is
specified and used.