test behavior of case search with multiple results
This commit adds a simple test to ensure we cover the
(expectedly) rare case when cases are found in multiple
backends. We can't currently cover the error cases well
because we silence all client errors. Ideally we'd test
this without error-silencing and using VCR, but that's
for a later iteration.