check for missing ZGW backend config in zgw_import_data command
Not having a ZGW backend configured is something that happens in
the wild, and more broadly, it would be a good thing if the
management command does not fail hard on a missing config (as now)
but gracefully exits with an error message. Better reporting to
users (including our Celery task wrapper for this command), and
cuts down on false-positive Sentry pings.