fix missing mock DigiD/eHerkenning backends in dev
We were too eager in 65304e2 in removing all the dev
backends, in an attempt to avoid divergence between
the base and dev authentication backends. Unfortunately,
we mistakenly also removed references to the mock backends,
which broke local development which suddenly required
production settings for DigiD and eHerkenning.

This commit makes more explicit that all we want to do on
dev is replace these backends whilst keeping the other
settings the same.