Disambiguate missing cache values from None
We have observed an issue with spurious cache hits in development,
that we've been unable to reproduce so far. Looking into this did
however prompt an audit of the caching decorator, which reveals
we potentially have an issue where a cached "None" value is mis-
interpreted as "not cached" rather than "cached as None". This
commit clarifies the expected behaviour with a test and a more
robust cache check.