Disable playwright browser cache in CI
As per the Playwright docs, caching browsers is not recommended, both
because downloading the binary is not much slower, but also because
system dependencies are required, and these cannot be cached. Even
worse, the system dependencies can become stale as the CI images
get upgraded. This seems to have happened recently, with persistent
"It looks like you are using Playwright Sync API inside the asyncio
loop." errors sporadically appearing (in this case, for webkit only)
and this being fixed, without further code changes, by disabling the
browser cache.