Update uWSGI settings for Docker
This commit limits the number of uWSGI settings that
are externally configurable. We are opinionated with
regards to certain settings (mainly to do with process
management), for others we provide sane defaults but
allow overrides if they're mainly to do with the
surrounding environment and resource constraints
(threads, processes, timeouts, etc.).