Make mozilla_django_oidc_db SessionRefresh middleware configurable
This is a workaround for an issue in mozilla_django_oidc_db which
hasn't yet been fully diagnosed, whereby the OIDC refresh middleware
errors when multiple OIDC providers are configured (e.g. for the admin.
and for DigiD). To limit the scope of the changes, we've made the
middleware configurable, but enabled by default so that any behavioral
changes are opt-in. This will have to be revisited once the issue
has been addressed. See also Taiga #3027.