[#3096] Add a vestiging field to the User model
This commit adds a `vestiging` attribute to the User model.
Vestigingen are for most use-cases, and in the eyes of most
clients, distinct users from the legal entity to which
the vestiging belongs. To handle this properly, we'll have
to make vestigingen first-class citizens, so we can e.g. store
profile information on a vestiging basis, and generally allow
vestigingen to interact with OIP on a full-fledged basis (e.g.
also having, for instance, different notification settings
then the main legal entity).