[#3101] Use User.vestiging within the OIDC eHerkenning backend
This commit uses the branch claim from the OIDC broker to authenticate
the user based on the vestiging (if present) or on the kvk alone (if
absent). It also persists some information into the session if a branch
restriction was present. This is because we have to assume, if such
a restriction was passed by the broker, the eHerkenning user is
restricted to only seeing that branch and no others (nor can it access
the user for the legal entity as such). When we implement the branch
selection, we'll have to retain this information to determine whether
a user is authorized to switch from one branch to another.