Use uv to compile dependencies
To preserve consistency with our previous setup using pip-compile,
we add uv.toml to exclude pip and setuptools from the output
of `uv pip compile` as they are considered unsafe. `--allow-unsafe`
will become the default in a future version of pip-compile. The
file can then be deleted (unless needed for other purposes).