[#3096] Make kvk attribute required for eherkenning users
The kvk attribute was practically already treated as a unique
identifier by the eHerkenning authentication backends, but this
was not yet enforced at the database level. In light of the upcoming
changes with regard to adding vestigingen to the user model, it makes
sense to tighten these constraints to make it less likely we'll run
into edge cases in narrowing down users to the correct kvk/vestiging
pair.