[#3096] Exclude tagged vestiging tests while we complete the feature
Adding vestigingen is a big change. To make this manageable,
we'll merge these changes to a feature branch, and to facilitate
parallel PRs, we'll tag the currently failing tests and fix them
as we go along, aiming to remove all tags and fix all tests
prior to merging the feature branch into the trunk.