Fix uv based compile dependencies by emitting base.txt
It seems the bundled pip-tools in uv does not by default output
the compiled requirements to a file if there is only a single
source, as with the first of the layered runs that is supposed to
emit base.txt. Because the --output-file parameter was not provided,
base.txt was not emitted and thus not picked up by the subsequent
runs which depend on this file. This commit adds the missing flag
to ensure base.txt is produced from base.in.