[#3097] Amend SAML eHerkenning backend to support User.vestiging
This commit amends the eHerkenning backend to resolve to either
the legal entity user (only kvk) or a specific branch (kvk and
vestiging), and persists any branch restrictions into the session
for use in the branch selection view. This should work identically
to the OIDC counterpart.

We only test the get_or_create_user method as the package is
vendored and we should be able to limit ourselves to verifying
the public API that we override returns the correct user for the
correct claims, without worrying about the low-level handling
(which is a package concern).