Specify expected Python version in compile_dependencies.sh
Without specifying the expected Python version in the
uv pip compile commands, uv will resort to discovering
the Python version, which could be non-deterministic,
depending on the the system on which the script is run.
This in turn means deps might be inadvertently removed
when running the script, as they may, for instance, only
be required for earlier Python versions.