Revert "[#3136] Add option to fetch only zaken where user has relevant role (#1676)"
This reverts commit 03910f2a335b79c0a832c273fb764101646aba16, reversing
changes made to a43af4dd04951562151ec19239e1b31aa36af2cc.

This PR has to be backported to 1.25, and upon asking around, the
best practice for backporting commits with migrations is to use the
forwardport pattern. Reverting this for now before it has been deployed
to any of the production environments.