Bump kombu to 5.5.1
This is to deal with the issue mentioned in:
https://taiga.maykinmedia.nl/project/dimpact-enschede-ssc-ict-1/issue/229

TLDR, Celery would sometimes be unable to reconnect to redis upon connection
loss. This will be fixed in Celery 5.5, but we can get an advance by
bumping Kombu to 5.5, which doesn't appear to have any breaking changes
relevant to our setup.