Merge on_kvk_change into account post_login sync
This commit consolidates the RSIN updating logic for the
post_login signal in the accounts module. Apart from being
cleaner (keeping user data synchronized is an accounts module
concern, not a kvk module concern), it also ensures all the
various eHerkenning user updates happen _before_ any pushes
to the klanten APIs take place. This was not guaranteed to
happen when the sync functions were split up, causing issues
in the klanten API syncs which, depending on the config,
might require the RSIN to be used to identify klanten.