Persist full session on eHerkenning user switch
When switching users for a different vestiging, the
user will initially have authenticated through SAML or
OIDC. In the latter case especially, there will be a
number of session variables on which the OIDC flows
depend. For instance, the ID token may be persisted
and this will be used for the logout flow.

When a user switches to a different vestiging-user, we
do not want to break any of these flows. Essentially, it
should be as if the session was not changed _except_ for the
user (and possible the branch selection flag).

To address this, we copy over all session variables, apart
from the Django auth session keys which obviously will change,
because we now have a new user.

The alternative was to copy over the specific variables which
the OIDC backend expects, but that leaves us vulnerable to
breaking interface changes in these packages. More generally,
other modules might depend on the existence of certain session
variables. In that light, simply copying over as much as we can
seems like the safest path forward.