[#3027] Add missing candidate backend check for CustomOIDCBackend
Because we have multiple OIDC backends configured, mozilla_django_oidc_db
relies on the _check_candidate_backend method to select the appropriate
backend for specific flows such as callbacks and, crucially, the
SessionRefresh middleware.

Our custom OIDC authentication backend, used for admin access via e.g.
Keycloak and Microsoft Entra, did not override the _check_candidate_backend
method and thus fell back on the default implementation, which simply
checked whether the config class was enabled or not.

In practice, this lead to a difficult to diagnose issue where by, as
a consequence of the SessionRefresh middleware, _any_ request would
be redirected back to the OIDC iDP after a configured interval
(15 minutes by default). However, if the user logged in with any
OIDC backend _other_ than the custom OIDC authentication backend,
the session would contain various params (state, ID tokens, and so
on) from that backend. This ultimately lead to mozzilla_django_oidc_db
incorrectly redirecting the user to the custom OIDC backend, which
would subsequently generate an error because the various parameters
were generated by a different backend.

This commit implements the missing override check, and also
reverts the workaround we introduced to allow the disabling of
this middleware, introduced in the v1.26.2 and v1.27.1 patch
releases, as the SessionRefresh middleware should now work as
expected.