Update VCR fixture naming for OpenKlant2 to be all lowercase
The default naming will in various places include cased string
fragments in the tests, and thus in the VCR fixtures. This is a
pain on case insensitive filesystems (e.g. on Mac), as git will
spuriously detect changed files when there are none there. This
commit ensures the generated names are all lower-case.