[#3128] Distinguish between company_name and branch_name
We update the `company_name` and the `branch_name` for eHerkenning users
on login. The `company_name` is the same for all branches. The
`branch_name` is retrieved from the vestigingsprofiel (if
user.vestiting is set) or set to the name of the hoofdvestiging (if
the user is logged in as "rechtspersoon" representing the company
as a whole).