[#3187] Ensure SSD view tests actually fetch XML and return PDF
It turns out our view tests were mocking the wrong methods, and thus
weren't actually returning rendered PDF files. This commit mocks the
correct methods and also uses a request mocker to stay closer to the
real-world flow. It is hard to assert against the content of a generated
PDF (we also can't use e.g. snapshot testing because the output is
indeterminate at a binary level for the same inputs). As something
is better than nothing, we at least assert we're getting actual
bytes and that these include the PDF magic number.