[#3188] Remove loon zorgverzekeringswet from SSD Jaaropgave
Remove the loon zorgverzekeringswet from the generated reports,
including the calculation logic. Some small visual hacks needed
to keep the layout consistent (using the style tag), but given
that this won't actually be user-facing but will simply be fed
into the PDF generation function, this shouldn't be too bad,
as we don't have to worry about browser behavior.