feat: disable the Django sites framework
The Django sites framework frequently causes deployment issues in new environments due to lacking a proper setup flow. We could obviously add that (django setup configuration has a built-in step for this), but given that nothing else uses the framework, it makes more sense just to disable it as a vestigial bit of functionality that we have no use for.

BREAKING CHANGE: Removes the Sites framework