feat: make primary klant backend optional
The primary klant backend could not be de-selected once selected.
Consequently, the contactform could not be removed from the footer; it
would always show up once a backend had been selected. The commit makes
the backend optional. If no primary klant backend is selected, the
contactform is absent from the footer.

Refs: Taiga issue #3167