feat: make SECRET_KEY_FALLBACKS environment configurable
Apart from allowing key rotation as a general best practice,
making the SECRET_KEY_FALLBACKS configurable can facilitate
instance migrations which keep existing sessions intact (using
the fallback) whilst progressively phasing these ciphertexts
out on the basis of the new primary SECRET_KEY.

We only allow a single value to avoid commas intended to be
part of a high-entropy SECRET_KEY being interpreted as a
list (and thereby giving several low-entropy secret keys).

This is in anticipation of some planned OIP migrations, to
give us the option to migrate crypto-verified state if we
need to.