feat: block eherkenning login in case of any KVK API errors
We block eherkenning login for eenmanszaken, who are required to log in
via digid. Instead of silencing errors from the KVK API (the status
quo), which has the effect that eenmanszaken can log in via eherkenning
after all, we check for this case and display an error message. This
has the effect that if there is an error when fetching the basisprofiel
of a company, login via eherkenning is down for everyone.

Refs: - https://taiga.maykinmedia.nl/project/open-inwoner/task/3042
      - https://taiga.maykinmedia.nl/project/open-inwoner/us/3039