feat: fix Service connected to CatalogusConfig on zgw_import_data
Previously, when syncing catalogus configs, we would match only
on URLs but not check the connected Service. This means that
certain CatalogusConfig objects may have gotten into a broken
state, due to existing prior to introducing the service attribute
on the CatalogusConfig model. In that case, there is a chance
the field was never properly updated, or that the value was set
to an incorrect service.

This is important becaus we use the Service to configure clients
for all nested types (zaaktypes, resultaattypes, etc.), and we might
get InvalidURL exceptions from ape_pie, because the api_root of the
client does not match that of the resource being fetched.