feat: update helptext for product button text
Updating the text for the CTABUTTON will only change the text of buttons
inside the contents of products. Buttons outside of the content have a
hardcoded (in the template) default of "Aanvraag starten". The help text
for the button text is updated to reflect the fact that changes only
affect buttons inside the contents of products.