fix: remove xml-related packages from Docker build
When libxml is installed (which it is through installing libxml2-dev),
then pip compiles lxml from source against the system libxml. This has
caused a mismatch between the system libxml available at runtime and the
libxml that lxml was compiled against during the build stage. Without
libxml on the system, pip falls back on pre-built binary wheels, which
have their own version of libxml built in.

Refs: https://taiga.maykinmedia.nl/project/maykin-intranet/issue/1169