feat: fetch and display taken with link to external form
API models for 'externe taken' are modified to reflect changes in
the API schema for tasks with external form; the plugin code for
rendering 'taken' is updated accordingly. Legacy models ('UrlTaak')
continue to be used for backward-compatibility. The config model
for the 'TasksPlugin' now has two fields for objecttype version:
one for the 'externformulier task', one for the old 'url task'

Bugfix: the objecttype UUID was incorrectly retrieved from the
'TasksConfig' class instead of an instance, which resulted in a
'DeferredAttribute' where we expect a string.

Refs: https://taiga.maykinmedia.nl/project/open-inwoner/us/3367