refactor: turn contact form into CMS page
The contact form is now a CMS page. The logic for handling form
submissions is kept inside the openklant "ContactFormView" to minimize
coupling with the CMS. The logic for checking if the contactform is
enabled and should be displayed is simplified: the old
`has_contactform_configuration` is removed, instead the more permissive
`contact_registration_enabled` is used everywhere.

Refs: https://taiga.maykinmedia.nl/project/open-inwoner/us/3406