feat: filter eHerkenning zaken by both KVK and vestigingsnummer
Zaken for eHerkenning users logged in as vestiging are now filtered by KVK
as well as vestigingsnummer. The filtering is done client-side as the
eSuite does not have a query parameter for this purpose.

For contactmomenten/questions, this filtering only works for OpenKlant
but not for eSuite as the latter does not support scoping of identifiers
(i.e. looking for a vestigingsnummer that's related to a KVK).

Refs: https://taiga.maykinmedia.nl/project/open-inwoner/us/3457