fix: add error msg for phonenumber DB constraint
Also removed the form-layer validation and error messages. We have
different login/registration flows and different flows; it is
cumbersome and error-prone to add the validation to some but not
others. The new approach is also in line with other constraint
checks for users, which are implemented at DB layer.

Refs: https://taiga.maykinmedia.nl/project/dimpact-enschede-ssc-ict-1/issue/345