tests: remove outgoing logger test
Because we're removing logging from CI, this test would not add much
with all the logging we have to add. It's also an isolated test that
does not add much information, we'd be better of directly asserting
that outgoing requests logs are enabled in the appropriate config.
For now, the need for cleaner test logs prevails.