feat: make hints in OIDC frontend logout optional via env
Because the ID token might include sensitive claims, and because these
might end-up in the querystring of client's user agent, this commit
adds an environment variable to optionally disable the hints when
redirecting the user to an OIDC frontend logout URL to avoid exposing
sensitive claims when needed. This can be required in PENTests for e.g.
DigiD.

The default remains to include the hints to avoid changing the
current behaviour.

The specs stipulate that the hints are optional, and if the token hint
is absent, the logout redirect must also be absent. In such cases,
the broker is supposed to prompt the user whether they indeed want
to logout (though we have seen brokers in the wild that treat the hints
as mandatory).

Fixes: Taiga story #3596